+++ extension-compose/src/main/java/com/mapbox/maps/extension/compose/internal/MapboxMapNode.kt
@@ -6,10 +6,24 @@ import androidx.compose.runtime.ComposeNode
 import androidx.compose.runtime.currentComposer
 import com.mapbox.common.Cancelable
 import com.mapbox.maps.CameraChangedCallback
+import com.mapbox.maps.MapIdleCallback
 import com.mapbox.maps.MapInitOptions
+import com.mapbox.maps.MapLoadedCallback
+import com.mapbox.maps.MapLoadingErrorCallback
 import com.mapbox.maps.MapView
 import com.mapbox.maps.MapboxExperimental
+import com.mapbox.maps.RenderFrameFinishedCallback
+import com.mapbox.maps.RenderFrameStartedCallback
+import com.mapbox.maps.ResourceRequestCallback
+import com.mapbox.maps.SourceAddedCallback
+import com.mapbox.maps.SourceDataLoadedCallback
+import com.mapbox.maps.SourceRemovedCallback
+import com.mapbox.maps.StyleDataLoadedCallback
+import com.mapbox.maps.StyleImageMissingCallback
+import com.mapbox.maps.StyleImageRemoveUnusedCallback
+import com.mapbox.maps.StyleLoadedCallback
 import com.mapbox.maps.extension.compose.DefaultSettingsProvider
+import com.mapbox.maps.extension.compose.MapEvents
 import com.mapbox.maps.extension.compose.animation.viewport.MapViewportState
 import com.mapbox.maps.plugin.attribution.attribution
 import com.mapbox.maps.plugin.attribution.generated.AttributionSettings
@@ -37,9 +51,11 @@ private class MapboxMapNode(
   val controller: MapView,
   initialClickListener: OnMapClickListener,
   initialLongClickListener: OnMapLongClickListener,
-  mapViewportState: MapViewportState
+  mapViewportState: MapViewportState,
+  mapEvents: MapEvents?,
 ) : MapNode() {
   private var cancelable: Cancelable? = null
+  private var mapEventCancelableHolder: MapEventCancelableHolder
   private val cameraChangedCallback = CameraChangedCallback {
     val cameraState = controller.getMapboxMap().cameraState
     mapViewportState.cameraState = cameraState
@@ -51,6 +67,7 @@ private class MapboxMapNode(
 
   init {
     mapViewportState.setMap(controller)
+    mapEventCancelableHolder = MapEventCancelableHolder(controller.getMapboxMap())
   }
 
   var mapViewportState = mapViewportState
@@ -78,6 +95,105 @@ private class MapboxMapNode(
       field = value
     }
 
+  var mapLoadedCallback: MapLoadedCallback? = mapEvents?.onMapLoaded
+    set(value) {
+      if (value == field) return
+      mapEventCancelableHolder.cancelAndSubscribeToMapLoaded(value)
+      field = value
+    }
+
+  var mapIdleCallback: MapIdleCallback? = mapEvents?.onMapIdle
+    set(value) {
+      if (value == field) return
+      mapEventCancelableHolder.cancelAndSubscribeToMapIdle(value)
+      field = value
+    }
+
+  var mapLoadingErrorCallback: MapLoadingErrorCallback? = mapEvents?.onMapLoadingError
+    set(value) {
+      if (value == field) return
+      mapEventCancelableHolder.cancelAndSubscribeToMapLoadingError(value)
+      field = value
+    }
+
+  var styleLoadedCallback: StyleLoadedCallback? = mapEvents?.onStyleLoaded
+    set(value) {
+      if (value == field) return
+      mapEventCancelableHolder.cancelAndSubscribeToStyleLoaded(value)
+      field = value
+    }
+
+  var styleDataLoadedCallback: StyleDataLoadedCallback? = mapEvents?.onStyleDataLoaded
+    set(value) {
+      if (value == field) return
+      mapEventCancelableHolder.cancelAndSubscribeToStyleDataLoaded(value)
+      field = value
+    }
+
+  var sourceDataLoadedCallback: SourceDataLoadedCallback? = mapEvents?.onSourceDataLoaded
+    set(value) {
+      if (value == field) return
+      mapEventCancelableHolder.cancelAndSubscribeToSourceDataLoaded(value)
+      field = value
+    }
+
+  var sourceAddedCallback: SourceAddedCallback? = mapEvents?.onSourceAdded
+    set(value) {
+      if (value == field) return
+      mapEventCancelableHolder.cancelAndSubscribeToSourceAdded(value)
+      field = value
+    }
+
+  var sourceRemovedCallback: SourceRemovedCallback? = mapEvents?.onSourceRemoved
+    set(value) {
+      if (value == field) return
+      mapEventCancelableHolder.cancelAndSubscribeToSourceRemoved(value)
+      field = value
+    }
+
+  var styleImageMissingCallback: StyleImageMissingCallback? = mapEvents?.onStyleImageMissing
+    set(value) {
+      if (value == field) return
+      mapEventCancelableHolder.cancelAndSubscribeToStyleImageMissing(value)
+      field = value
+    }
+
+  var styleImageRemoveUnusedCallback: StyleImageRemoveUnusedCallback? =
+    mapEvents?.onStyleImageRemoveUnused
+    set(value) {
+      if (value == field) return
+      mapEventCancelableHolder.cancelAndSubscribeToStyleImageRemoveUnused(value)
+      field = value
+    }
+
+  var onCameraChangedCallback: CameraChangedCallback? = mapEvents?.onCameraChanged
+    set(value) {
+      if (value == field) return
+      mapEventCancelableHolder.cancelAndSubscribeToCameraChanged(value)
+      field = value
+    }
+
+  var renderFrameStartedCallback: RenderFrameStartedCallback? = mapEvents?.onRenderFrameStarted
+    set(value) {
+      if (value == field) return
+      mapEventCancelableHolder.cancelAndSubscribeToRenderFrameStarted(value)
+      field = value
+    }
+
+  var renderFrameFinishedCallback: RenderFrameFinishedCallback? = mapEvents?.onRenderFrameFinished
+    set(value) {
+      if (value == field) return
+      mapEventCancelableHolder.cancelAndSubscribeToRenderFrameFinished(value)
+      field = value
+    }
+
+  var resourceRequestCallback: ResourceRequestCallback? = mapEvents?.onResourceRequest
+    set(value) {
+      if (value == field) return
+      mapEventCancelableHolder.cancelAndSubscribeToResourceRequest(value)
+      field = value
+    }
+
   override fun onAttached() {
     controller.gestures.apply {
       addNonDefaultOnClickListener(clickListener)
@@ -86,6 +202,22 @@ private class MapboxMapNode(
     mapViewportState.cameraState = controller.getMapboxMap().cameraState
     controller.viewport.addStatusObserver(viewportStatusObserver)
     cancelable = controller.getMapboxMap().subscribeCameraChanged(cameraChangedCallback)
+    mapEventCancelableHolder.apply {
+      cancelAndSubscribeToMapLoaded(mapLoadedCallback)
+      cancelAndSubscribeToMapIdle(mapIdleCallback)
+      cancelAndSubscribeToMapLoadingError(mapLoadingErrorCallback)
+      cancelAndSubscribeToStyleLoaded(styleLoadedCallback)
+      cancelAndSubscribeToStyleDataLoaded(styleDataLoadedCallback)
+      cancelAndSubscribeToSourceDataLoaded(sourceDataLoadedCallback)
+      cancelAndSubscribeToSourceAdded(sourceAddedCallback)
+      cancelAndSubscribeToSourceRemoved(sourceRemovedCallback)
+      cancelAndSubscribeToStyleImageMissing(styleImageMissingCallback)
+      cancelAndSubscribeToStyleImageRemoveUnused(styleImageRemoveUnusedCallback)
+      cancelAndSubscribeToCameraChanged(onCameraChangedCallback)
+      cancelAndSubscribeToRenderFrameStarted(renderFrameStartedCallback)
+      cancelAndSubscribeToRenderFrameFinished(renderFrameFinishedCallback)
+      cancelAndSubscribeToResourceRequest(resourceRequestCallback)
+    }
   }
 
   override fun onRemoved() {
@@ -102,6 +234,7 @@ private class MapboxMapNode(
       removeNonDefaultOnLongClickListener(longClickListener)
     }
     cancelable?.cancel()
+    mapEventCancelableHolder.cancelAll()
     controller.viewport.removeStatusObserver(viewportStatusObserver)
   }
 }
@@ -119,6 +252,7 @@ internal fun MapboxMapComposeNode(
   mapViewportState: MapViewportState,
   onMapClickListener: OnMapClickListener,
   onMapLongClickListener: OnMapLongClickListener,
+  mapEvents: MapEvents?,
 ) {
   val mapApplier = currentComposer.applier as MapApplier
   ComposeNode<MapboxMapNode, MapApplier>(
@@ -127,7 +261,8 @@ internal fun MapboxMapComposeNode(
         mapApplier.mapView,
         onMapClickListener,
         onMapLongClickListener,
-        mapViewportState
+        mapViewportState,
+        mapEvents,
       )
     },
     update = {
@@ -162,6 +297,22 @@ internal fun MapboxMapComposeNode(
       }
       update(onMapLongClickListener) { listener ->
         this.longClickListener = listener
+      }
+      update(mapEvents) {
+        this.mapLoadedCallback = mapEvents?.onMapLoaded
+        this.mapIdleCallback = mapEvents?.onMapIdle
+        this.mapLoadingErrorCallback = mapEvents?.onMapLoadingError
+        this.styleLoadedCallback = mapEvents?.onStyleLoaded
+        this.styleDataLoadedCallback = mapEvents?.onStyleDataLoaded
+        this.sourceDataLoadedCallback = mapEvents?.onSourceDataLoaded
+        this.sourceAddedCallback = mapEvents?.onSourceAdded
+        this.sourceRemovedCallback = mapEvents?.onSourceRemoved
+        this.styleImageMissingCallback = mapEvents?.onStyleImageMissing
+        this.styleImageRemoveUnusedCallback = mapEvents?.onStyleImageRemoveUnused
+        this.onCameraChangedCallback = mapEvents?.onCameraChanged
+        this.renderFrameStartedCallback = mapEvents?.onRenderFrameStarted
+        this.renderFrameFinishedCallback = mapEvents?.onRenderFrameFinished
+        this.resourceRequestCallback = mapEvents?.onResourceRequest
       }
     }
   )